LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-label-element/proxy-modifier-click-to-associated-element.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "key".
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "key".
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "key".

Harness Error (FAIL), message = Unhandled rejection: Unknown source type "key".

PASS label with for attribute should proxy click events to the associated element on click
TIMEOUT label with for attribute should proxy click events to the associated element on shift click Test timed out
TIMEOUT label with for attribute should proxy click events to the associated element on alt click Test timed out
TIMEOUT label with for attribute should proxy click events to the associated element on meta click Test timed out

Harness Error (FAIL), message = Unhandled rejection: Unknown source type "key".

PASS label with for attribute should proxy click events to the associated element on click
TIMEOUT label with for attribute should proxy click events to the associated element on shift click Test timed out
TIMEOUT label with for attribute should proxy click events to the associated element on alt click Test timed out
TIMEOUT label with for attribute should proxy click events to the associated element on meta click Test timed out
foo foo foo foo
